feat: 카트 api 및 주문로직 테스트코드 작성 및 수정

+++ src/main/java/com/example/outsourcing/domain/shop/dto/request/ShopRequestDto.java
@@ -26,5 +26,5 @@ public class ShopRequestDto {
 
     private LocalTime openAt;
     private LocalTime closeAt;
-    private Double minPrice;
+    private Integer minPrice;
 }
